requirements.txt
discord.py==2.3.2
pylint==3.3.4